Use the same kitchen sink test data for all test projects
