+++ common/tests/codegen-tests/playwright-csharp-nunit/playwright-csharp-nunit.test.ts
@@ -4,8 +4,37 @@ import path from "path";
 import { IProgressEvent } from "../../../src/ipc-defs";
 import { generateCode } from "../../../src/codegen";
 import { writeOutputProjectFiles, createRmTestProject } from "../../test-helpers/rm-project-generator";
-import { createPlaywrightNunitTestData } from "./playwright-csharp-nunit.test-data";
 import { createTempDir } from "../../test-helpers/fsHelpers";
+
+import { AutomationFramework, Indent, Language, StandardFolder, TestFramework } from "../../../src/file-defs";
+import { RmpSpec } from "../../test-helpers/rm-project-spec.types";
+import { createTestDataKitchenSink } from "../playwright-typescript/playwright-typescript-kitchen-sink.test-data";
+
+export const createPlaywrightNunitTestData = (): RmpSpec => {
+  const kitchenSinkData = createTestDataKitchenSink();
+  return {
+    projectName: "google-test-playwright-nunit",
+    content: {
+      fileVersion: 1,
+      name: "",
+      description: "",
+      automationFramework: AutomationFramework.Playwright,
+      testFramework: TestFramework.NUnit,
+      language: Language.CSharp,
+      rootNamespace: "",
+      indent: Indent.Spaces,
+      indentSize: 4,
+      testIdAttributeName: "",
+    },
+
+    configFiles: kitchenSinkData.configFiles,
+    pages: kitchenSinkData.pages,
+    testcases: kitchenSinkData.testcases,
+    testroutines: kitchenSinkData.testroutines,
+    testsuites: kitchenSinkData.testsuites,
+    outputFiles: [],
+  };
+};
 
 test("CodeGen Playwright CSharp Nunit - General", async () => {
   // Arrange
